LayoutTests/imported/w3c/web-platform-tests/css/css-align/gaps/gap-animation-001-expected.txt


FAIL gap is interpolable assert_equals: expected "50px" but got "50px 50px"
